toy-cdk/requirements.txt
aws-cdk.aws-lambda-python==1.86.0
